start moving newtaskPopup to ts
