+++ tasks/models.py
@@ -1,7 +1,7 @@
 from django.db import models
 import uuid
 from django.utils import timezone
-
+from django.core.validators import RegexValidator
 
 
 class Task(models.Model):
@@ -17,6 +17,10 @@ class Task(models.Model):
     due = models.DateTimeField(help_text="Enter the date the task is due", default=timezone.now().replace(hour=23, minute=59, second=59, microsecond=999999))
     complete = models.BooleanField(default=False)
     completed_at = models.DateTimeField(editable=False, blank=True, null=True)
+    color = models.CharField(max_length=7, editable=True, blank=True, null=True, default="#ffffff", validators=[RegexValidator(
+        regex=r'^#(?:[0-9a-fA-F]{3}){1,2}$',
+        message='Enter a valid hex code'
+    )])
 
     # TODO null for user should not be true
     # user =  models.ForeignKey('User', null=True)
@@ -35,7 +39,6 @@ class Task(models.Model):
     def __str__(self):
         """String for representing the MyModelName object (in Admin site etc.)."""
         return f'{self.id}: {self.title}'
-
 
     # def get_absolute_url(self):
     #     """Returns the url to access a particular instance of the model."""
