+++ tasks/serializers.py
@@ -6,4 +6,4 @@ class TaskSerializer(serializers.ModelSerializer):
 
     class Meta:
         model = Task 
-        fields = ('id', 'title', 'complete', 'completed_at', 'start', 'due', 'description', 'created_at', 'updated_at')
+        fields = ('id', 'title', 'complete', 'completed_at', 'start', 'due', 'description', 'created_at', 'updated_at', 'color')
